+++ jest.config.ts
@@ -13,15 +13,9 @@ const config: Config = {
   },
   collectCoverageFrom: ['**/*.(t|j)s'],
   coverageDirectory: './coverage',
-  globals: {
-    /**
-     * Because @meeco/sdk expects it...
-     */
-    FormData: null,
-  },
   coveragePathIgnorePatterns: ['/node_modules/', '/src/test-utils/'],
   moduleNameMapper: {
-    '^@sd-jwt/test-utils(.*)$': '<rootDir>/src/test-utils/$1',
+    '^(\\.\\/.+)\\.js$': '$1',
   },
 };
