+++ package.json
@@ -1,7 +1,7 @@
 {
-  "name": "sd-jwt-vc",
+  "name": "@meeco/sd-jwt-vc",
   "version": "0.0.1",
-  "description": "SD-JWT implementation in typescript",
+  "description": "SD-JWT VC implementation in typescript",
   "source": "src/index.ts",
   "private": true,
   "scripts": {
@@ -18,7 +18,8 @@
     "test": "jest --verbose"
   },
   "engines": {
-    "node": ">=18"
+    "node": ">=18",
+    "npm": ">=8.0.0"
   },
   "repository": "@meeco/sd-jwt-vc",
   "keywords": [
@@ -38,23 +39,31 @@
   "main": "./dist/node/cjs/index.js",
   "browser": "./dist/browser/index.js",
   "types": "./dist/types/index.d.ts",
+  "files": [
+    "dist/**/package.json",
+    "dist/**/*.js",
+    "dist/types/**/*.d.ts",
+    "!dist/**/*.bundle.js",
+    "!dist/**/*.umd.js",
+    "!dist/**/*.min.js"
+  ],
   "author": "Meeco",
   "devDependencies": {
-    "@types/jest": "^29.5.4",
-    "@typescript-eslint/eslint-plugin": "^6.7.0",
-    "@typescript-eslint/parser": "^6.7.0",
+    "@types/jest": "^29.5.5",
+    "@typescript-eslint/eslint-plugin": "^6.7.3",
+    "@typescript-eslint/parser": "^6.7.3",
     "esbuild": "^0.19.3",
-    "eslint": "^8.49.0",
+    "eslint": "^8.50.0",
     "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-prettier": "^5.0.0",
-    "jest": "^29.6.4",
+    "jest": "^29.7.0",
+    "jose": "^4.14.6",
     "npm-run-all": "^4.1.5",
     "prettier": "^3.0.3",
     "ts-jest": "^29.1.1",
     "ts-node": "^10.9.1",
     "tsconfig-paths": "^4.2.0",
-    "typescript": "^5.2.2",
-    "jose": "^4.14.6"
+    "typescript": "^5.2.2"
   },
   "dependencies": {
     "sd-jwt": "file:../sd-jwt"
