+++ src/index.ts
@@ -1,3 +1,3 @@
-export * from './holder';
-export * from './issuer';
-export * from './types';
+export * from './holder.js';
+export * from './issuer.js';
+export * from './types.js';
