+++ src/test-utils/helpers.ts
@@ -1,6 +1,8 @@
+import { createHash, randomBytes } from 'crypto';
 import { JWK, JWTHeaderParameters, JWTPayload, KeyLike, SignJWT, importJWK, jwtVerify } from 'jose';
-import { KeyBindingVerifier, Signer, Verifier, decodeJWT } from 'sd-jwt';
-import { supportedAlgorithm } from '../util';
+import { Hasher, KeyBindingVerifier, Signer, Verifier, base64encode, decodeJWT } from 'sd-jwt';
+import { NonceGenerator } from '../types';
+import { defaultHashAlgorithm, supportedAlgorithm } from '../util';
 
 export function signerCallbackFn(privateKey: Uint8Array | KeyLike): Signer {
   return (protectedHeader: JWTHeaderParameters, payload: JWTPayload): Promise<string> => {
@@ -50,4 +52,28 @@ export function verifierCallbackFn(publicKey: Uint8Array | KeyLike): Verifier {
     const verifiedKbJWT = await jwtVerify(jwt, publicKey);
     return !!verifiedKbJWT;
   };
+}
+
+/**
+ *
+ * The `algorithm` is dependent on the available algorithms supported by the
+ * version of OpenSSL on the platform. Examples are `'sha256'`, `'sha512'`, etc.
+ * On recent releases of OpenSSL, `openssl list -digest-algorithms` will
+ * display the available digest algorithms.
+ * @param algo
+ * @returns
+ */
+export function hasherCallbackFn(alg: string = defaultHashAlgorithm): Hasher {
+  return (data: string): string => {
+    const digest = createHash(alg).update(data).digest();
+    return base64encode(digest);
+  };
+}
+
+export function nonceGeneratorCallbackFn(length: number = 16): NonceGenerator {
+  return () => generateNonce(length);
+}
+
+export function generateNonce(length: number = 16): string {
+  return randomBytes(length).toString('base64');
 }
