+++ src/util.ts
@@ -1,8 +1,3 @@
-import { createHash, randomBytes } from 'crypto';
-
-import { Hasher, base64encode } from 'sd-jwt';
-import { NonceGenerator } from './types';
-
 export enum supportedAlgorithm {
   EdDSA = 'EdDSA',
   ES256 = 'ES256',
@@ -26,28 +21,4 @@ export function isValidUrl(url: string): boolean {
   } catch {
     return false;
   }
-}
-
-/**
- *
- * The `algorithm` is dependent on the available algorithms supported by the
- * version of OpenSSL on the platform. Examples are `'sha256'`, `'sha512'`, etc.
- * On recent releases of OpenSSL, `openssl list -digest-algorithms` will
- * display the available digest algorithms.
- * @param algo
- * @returns
- */
-export function hasherCallbackFn(alg: string = defaultHashAlgorithm): Hasher {
-  return (data: string): string => {
-    const digest = createHash(alg).update(data).digest();
-    return base64encode(digest);
-  };
-}
-
-export function nonceGeneratorCallbackFn(length: number = 16): NonceGenerator {
-  return () => generateNonce(length);
-}
-
-export function generateNonce(length: number = 16): string {
-  return randomBytes(length).toString('base64');
 }
