+++ src/verifier.spec.ts
@@ -1,7 +1,7 @@
 import { JWK, importJWK } from 'jose';
 import { decodeJWT } from 'sd-jwt';
-import { kbVeriferCallbackFn, verifierCallbackFn } from './test-utils/helpers';
-import { defaultHashAlgorithm, hasherCallbackFn } from './util';
+import { hasherCallbackFn, kbVeriferCallbackFn, verifierCallbackFn } from './test-utils/helpers';
+import { defaultHashAlgorithm } from './util';
 import { Verifier } from './verifier';
 
 describe('Verifier', () => {
