+++ tsconfig/browser.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../tsconfig.json",
+  "files": ["../src/index.ts"],
+  "compilerOptions": {
+    "declaration": false,
+    "target": "ES2018",
+    "module": "ES6",
+    "outDir": "../dist/browser",
+    "moduleResolution": "node"
+  }
+}
