+++ tsconfig/node-cjs.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../tsconfig.json",
+  "files": ["../src/index.ts"],
+  "compilerOptions": {
+    "types": ["node"],
+    "target": "ES2019",
+    "module": "CommonJS",
+    "outDir": "../dist/node/cjs",
+    "declaration": false
+  }
+}
