+++ tsconfig/node-esm.json
@@ -1,0 +1,10 @@
+{
+  "extends": "../tsconfig.json",
+  "files": ["../src/index.ts"],
+  "compilerOptions": {
+    "outDir": "../dist/node/esm",
+    "declaration": false,
+    "module": "ES2020",
+    "moduleResolution": "node"
+  }
+}
