+++ tsconfig/types.json
@@ -1,0 +1,12 @@
+{
+  "extends": "../tsconfig.json",
+  "files": ["../src/index.ts"],
+  "compilerOptions": {
+    "module": "nodenext",
+    "outDir": "../dist/types",
+    "declaration": true,
+    "emitDeclarationOnly": true,
+    "removeComments": false,
+    "moduleResolution": "nodenext"
+  }
+}
